Sampling all the available frequencies
Now the code does the routine for all the sets of frequencies available.
The code also compares the radius obtained from the 3 frequencies. If the ratio of two radii is within 1e-4 of 1 then we accpect the mass and spin value.
Now since we have 2 such pairs we can see if there is an agreement between 3 of the radius.
And that mass and spin pair when used across all the frequencies produces consistent radius from the 3 freqs

Future trial, Use the known mass estimate and try to see if something else is better constrained