Using all three obtained radius to get allowed m&a
In this version, all 3 obtained radii are used to constrain the mass and spin.
The ratio of the radii if lies within a particular interval around 1 will be accepted as a possible soln and
the mass and spin required to make this happen will be considered.

Plotting of the possible mass spin pairs is also done.

Next version will use the error in the freq to get the radius interval. If the radius intervals overlap then the mass spin pair can be accepted upto 90% confidence