Major update - working code for all mass spin and freqs
The current version of code makes a grid of mass and spin and runs for different frequencies to determine the radius.
It uses the error on the freq to get a range on the radius. If the 3 radii obtained have a common interval then the mass spin pair is accepted.

The code writes out the accepted mass spin pairs into a text file for all the observations it works with.
The multiple text files can be later used to produce the plot of mass spin distribution

The next commit addresses this code.