Modified the code to give out the arrays of radii
Now the code returns a list of 4*len(mass_guess)*len(spin_guess) numbers which correspond to the computed radii and the flag which indicates if the mass spin combo is selected or not

Currently the radius information is not save anywhere. It has been made accessible post the parallelisation run.